performance: relocate pragma callback, improve comments
